requirements.txt
Flask==2.3.3
pyserial==3.5 